Added possibility to forward clientX and clientY values
When dispatching a new long-press event, none of values from the original event was added to the new long-press event.

I've updated the code, such that clientX and clientY are added to the detail parameter, such that they can be used to position eg. a popup against.

Besides that, I've added a test, to ensure that it works.